Keep executor tests within package size limits
